Configure `start` command to use / as root
The `start` command in the package.json file
created the application expecting assets
to be served at /dota2quiz instead of the root,
which meant that images were not loading properly when
developing on local.

Root cause is the `hompage: '.'` setting in the package.json
file, and the solution is to either change that or to
prefix the run command with a command to set the `PUBLIC_URL`
environmental variable.